+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/service/ClusterNetworkAddressConfigProvider.java
@@ -22,7 +22,7 @@ public interface ClusterNetworkAddressConfigProvider {
     HostPort getClusterBootstrapAddress();
 
     /**
-     * Address of broker with the given node id, includes the port. Note that
+     * Address of broker with the given node id, (advertised hostname and bind port). Note that
      * {@code nodeId} are generally expected to be consecutively numbered and starting from zero. However,
      * gaps in the sequence can potentially emerge as the target cluster topology evolves.
      *
@@ -33,10 +33,22 @@ public interface ClusterNetworkAddressConfigProvider {
     HostPort getBrokerAddress(int nodeId) throws IllegalArgumentException;
 
     /**
-     * Generates the node id implied by the given broker address. This method make sense only for
-     * implementation that embed node id information into the broker address.  This information is
-     * used at startup time to allow a client that already in possession of a broker address
-     * to reconnect to the cluster via Kroxylicious using only that address.
+     * Advertised address of broker with the given node id, (advertised hostname and advertised port). This is
+     * what is returned to clients and may differ from the node's bind port as presented by {@link #getBrokerAddress(int)}.
+     * This enables Kroxylicious to sit behind yet another proxy that uses a different port from the kroxylicious bind port.
+     * @param nodeId node id
+     * @return the port to advertise for the nodeId
+     * @throws IllegalArgumentException if this provider cannot produce a broker address for the given nodeId.
+     */
+    default HostPort getAdvertisedBrokerAddress(int nodeId) throws IllegalArgumentException {
+        return getBrokerAddress(nodeId);
+    }
+
+    /**
+     * Generates the node id implied by the given broker address (advertised hostname and bind port).
+     * This method make sense only for implementation that embed node id information into the broker
+     * address.  This information is used at startup time to allow a client that already in possession
+     * of a broker address to reconnect to the cluster via Kroxylicious using only that address.
      * <br/>
      * This is an optional method. An implementation can return null.
      *